抽取数据库操作
将数据库操作抽取为DAO层，方便以后扩展
